+++ electron/preload.cjs
@@ -19,6 +19,7 @@ contextBridge.exposeInMainWorld("mailAPI", {
   },
   mail: {
     setFlag: (params) => ipcRenderer.invoke("mail:setFlag", params),
+    fetchBody: (params) => ipcRenderer.invoke("mail:fetchBody", params),
   },
   cache: {
     read: (params) => ipcRenderer.invoke("cache:read", params),